Fixing defaultOption directive for ngOptions
defaultOption did not work correctly with ng-options directive, because
it had priority set to a negative value. ngOptions directive has a
'terminal' property defined in its directive definition object, and
because of that defaultOption directive was never linked. Current case
works well with cases when ngOptions is used, as well as with cases when
user defines the options using HTML tags.